+++ src/app/api/links/list/route.ts
@@ -77,7 +77,7 @@ export async function GET(req: Request) {
                 totalPages: totalPages,
                 currentPage: currentPage,
                 pageSize: pageSize
-            }, allowEdit: userPlan, allowDA: userPlan
+            }, allowEdit: userPlan, allowDA: userPlan, allowQRCode: userPlan
         }, { status: 200 });
     } catch (error) {
         console.error("Error fetching list:", error);