+++ src/app/login/page.tsx
@@ -39,14 +39,14 @@ export default function Page() {
     if (errors.email) {
       toast.error("Please enter a valid email", {
         duration: 3000,
-        position: "top-center",
+        position: "bottom-center",
         icon: "🚫",
         style: { backgroundColor: "#790000", color: "#fff" },
       });
     } else if (errors.password) {
       toast.error(errors.password.message || "Password is required", {
         duration: 3000,
-        position: "top-center",
+        position: "bottom-center",
         icon: "🚫",
         style: { backgroundColor: "#790000", color: "#fff" },
       });
@@ -81,7 +81,7 @@ export default function Page() {
 
       toast.success("Login successful, redirecting...", {
         duration: 5000,
-        position: "top-center",
+        position: "bottom-center",
         icon: "🚀",
         style: { backgroundColor: "#005f08", color: "#fff" },
       });
@@ -94,7 +94,7 @@ export default function Page() {
       console.error("Login error:", error);
       toast.error(error.message, {
         duration: 5000,
-        position: "top-center",
+        position: "bottom-center",
         icon: "🚫",
         style: { backgroundColor: "#790000", color: "#fff" },
       });